integration/shared-auth/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
passlib[bcrypt]>=1.7.4
asyncpg>=0.29.0
